move resolve_pending to its own file
