+++ src/eth/storage/resolve_pending.rs
@@ -1,0 +1,237 @@
+use parking_lot::RwLockReadGuard;
+
+use crate::eth::primitives::Account;
+use crate::eth::primitives::BlockNumber;
+use crate::eth::primitives::PointInTime;
+use crate::eth::primitives::Slot;
+use crate::eth::primitives::StorageError;
+use crate::eth::storage::ReadKind;
+use crate::eth::storage::StratusStorage;
+use crate::eth::storage::TxCount;
+use crate::eth::storage::stratus_storage::EntityRead;
+use crate::infra::metrics::MetricLabelValue;
+
+/// Prevents construction of [`MinedPointInTime`] outside this module.
+/// `Seal` is public (so the enum variants can be pattern-matched) but cannot be constructed
+/// externally because its field type is private.
+#[derive(Debug)]
+pub struct Seal(SealPrivate);
+
+#[derive(Debug)]
+struct SealPrivate;
+
+/// A [`PointInTime`] that has been resolved past the pending case.
+///
+/// `Latest` carries an optional `transient_state_lock` read guard held across the latest read.
+///
+/// The [`Seal`] field makes both variants impossible to construct outside this module,
+/// while still allowing pattern matching externally.
+#[derive(Debug, strum::Display)]
+pub enum MinedPointInTime<'a> {
+    #[strum(to_string = "latest")]
+    Latest(Seal, Option<RwLockReadGuard<'a, ()>>),
+    #[strum(to_string = "past")]
+    Past(Seal, BlockNumber),
+}
+
+impl<'a> MinedPointInTime<'a> {
+    fn mined(guard: Option<RwLockReadGuard<'a, ()>>) -> Self {
+        Self::Latest(Seal(SealPrivate), guard)
+    }
+
+    fn mined_past(number: BlockNumber) -> Self {
+        Self::Past(Seal(SealPrivate), number)
+    }
+
+    /// Extracts the read guard if present, leaving `Mined(None)` in its place.
+    fn take_guard(&mut self) -> Option<RwLockReadGuard<'a, ()>> {
+        match self {
+            Self::Latest(_, guard) => guard.take(),
+            Self::Past(_, _) => None,
+        }
+    }
+}
+
+impl From<MinedPointInTime<'_>> for MetricLabelValue {
+    fn from(value: MinedPointInTime<'_>) -> Self {
+        Self::Some(value.to_string())
+    }
+}
+
+/// Unlocks the guard fairly when dropped.
+impl<'a> Drop for MinedPointInTime<'a> {
+    fn drop(&mut self) {
+        if let Some(guard) = self.take_guard() {
+            RwLockReadGuard::unlock_fair(guard);
+        }
+    }
+}
+
+/// Outcome of resolving pending state for a read.
+#[derive(Debug)]
+pub(super) enum Resolved<'a, T> {
+    /// Found in the pending cache.
+    PendingCache(T),
+    /// Found in temporary storage.
+    Temp(T),
+    /// Nothing pending.
+    Miss(MinedPointInTime<'a>),
+}
+
+/// Pending-state resolution, generic over the entity being read.
+pub(super) trait Resolve: EntityRead {
+    fn resolve(s: &StratusStorage, pit: PointInTime, key: Self::Key, kind: ReadKind) -> Result<Resolved<'_, Self>, StorageError> {
+        if pit == PointInTime::Pending {
+            if matches!(kind, ReadKind::Transaction)
+                && let Some(value) = Self::read_pending_cache(s, key)
+            {
+                return Ok(Resolved::PendingCache(value));
+            }
+            if let Some(value) = Self::read_temp(s, key, kind)? {
+                return Ok(Resolved::Temp(value));
+            }
+        }
+        Ok(Resolved::Miss(s.resolve_mined_point(pit, kind)))
+    }
+}
+
+impl Resolve for Account {}
+
+impl Resolve for Slot {}
+
+impl StratusStorage {
+    /// Determines the mined point-in-time for a read.
+    fn resolve_mined_point(&self, pit: PointInTime, kind: ReadKind) -> MinedPointInTime<'_> {
+        if let PointInTime::MinedPast(number) = pit {
+            return MinedPointInTime::mined_past(number);
+        }
+        match kind {
+            ReadKind::Call((block_number, tx_count)) => {
+                let guard = self.transient_state_lock.read();
+                let mined = self.read_mined_block_number();
+                if (block_number, tx_count) >= (mined, TxCount::Full) {
+                    MinedPointInTime::mined(Some(guard))
+                } else {
+                    drop(guard);
+                    let target = match tx_count {
+                        TxCount::Partial(_) => block_number.prev().unwrap_or_default(),
+                        TxCount::Full => block_number,
+                    };
+                    MinedPointInTime::mined_past(target)
+                }
+            }
+            ReadKind::Transaction | ReadKind::RPC => MinedPointInTime::mined(None),
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::super::StratusStorage;
+    use super::Resolve;
+    use crate::eth::primitives::Address;
+    use crate::eth::primitives::BlockNumber;
+    use crate::eth::primitives::PointInTime;
+    use crate::eth::primitives::Slot;
+    use crate::eth::primitives::SlotIndex;
+    use crate::eth::storage::ReadKind;
+    use crate::eth::storage::TxCount;
+
+    #[test]
+    fn pending_partial_call_latest_becomes_stale_once_block_is_mined() {
+        let storage = StratusStorage::new_test().expect("failed to build test storage");
+
+        let address = Address::ZERO;
+        let index = SlotIndex::ZERO;
+
+        // Pending call: block_number is the pending block (mined + 1 = 6), pinned to tx 0.
+        let mined_at_start = 5u64;
+        let pending_block_number = BlockNumber::from(mined_at_start + 1);
+        storage.set_mined_block_number(BlockNumber::from(mined_at_start));
+
+        let kind = ReadKind::Call((pending_block_number, TxCount::Partial(0)));
+
+        // At call start: block 6 is still pending (mined=5). Latest (block 5) is a safe base.
+        // resolve_slot should return Miss(Mined(Some(guard))).
+        let resolved = Slot::resolve(&storage, PointInTime::Pending, (address, index), kind).expect("resolve_slot");
+        match resolved {
+            super::Resolved::Miss(mut point) => {
+                assert!(
+                    matches!(point, super::MinedPointInTime::Latest(_, _)),
+                    "should read latest mined base while block is still pending"
+                );
+                assert!(point.take_guard().is_some(), "guard should be held for valid latest read");
+            }
+            other => panic!("expected Miss, got {other:?}"),
+        }
+
+        // The pending block is mined mid-call, advancing the mined tip to 6.
+        storage.set_mined_block_number(pending_block_number);
+
+        // The call is now stale: reading "latest" would observe block 6's aggregate state,
+        // not the tx-0 base. resolve_slot should downgrade to MinedPast(5) = b.prev() with no guard.
+        let resolved = Slot::resolve(&storage, PointInTime::Pending, (address, index), kind).expect("resolve_slot");
+        match resolved {
+            super::Resolved::Miss(mut point) => {
+                assert!(!matches!(point, super::MinedPointInTime::Latest(_, _)), "stale call should not read latest");
+                match &point {
+                    super::MinedPointInTime::Past(_, number) => {
+                        assert_eq!(
+                            *number,
+                            BlockNumber::from(mined_at_start),
+                            "stale Partial call should downgrade to MinedPast(b.prev())"
+                        );
+                    }
+                    other => panic!("expected Past, got {other:?}"),
+                }
+                assert!(point.take_guard().is_none(), "no guard for historical read");
+            }
+            other => panic!("expected Miss, got {other:?}"),
+        }
+    }
+
+    #[test]
+    fn mined_full_call_downgrades_to_minedpast_block_not_prev() {
+        let storage = StratusStorage::new_test().expect("failed to build test storage");
+
+        let address = Address::ZERO;
+        let index = SlotIndex::ZERO;
+
+        // Mined Full call: block_number = 5, mined = 5 → valid (b >= mined).
+        let call_block = BlockNumber::from(5u64);
+        storage.set_mined_block_number(call_block);
+
+        let kind = ReadKind::Call((call_block, TxCount::Full));
+
+        let resolved = Slot::resolve(&storage, PointInTime::Mined, (address, index), kind).expect("resolve_slot");
+        match resolved {
+            super::Resolved::Miss(mut point) => {
+                assert!(
+                    matches!(point, super::MinedPointInTime::Latest(_, _)),
+                    "Full call should read latest while block is the mined tip"
+                );
+                assert!(point.take_guard().is_some(), "guard should be held for valid latest read");
+            }
+            other => panic!("expected Miss, got {other:?}"),
+        }
+
+        // A newer block is mined mid-call, advancing the mined tip to 6.
+        storage.set_mined_block_number(BlockNumber::from(6u64));
+
+        // Stale: b=5 < mined=6. Full → MinedPast(5), NOT MinedPast(4).
+        let resolved = Slot::resolve(&storage, PointInTime::Mined, (address, index), kind).expect("resolve_slot");
+        match resolved {
+            super::Resolved::Miss(mut point) => {
+                assert!(!matches!(point, super::MinedPointInTime::Latest(_, _)), "stale call should not read latest");
+                match &point {
+                    super::MinedPointInTime::Past(_, number) => {
+                        assert_eq!(*number, call_block, "stale Full call should downgrade to MinedPast(block_number), not prev()");
+                    }
+                    other => panic!("expected Past, got {other:?}"),
+                }
+                assert!(point.take_guard().is_none(), "no guard for historical read");
+            }
+            other => panic!("expected Miss, got {other:?}"),
+        }
+    }
+}
